# apps/account/internal/core/models/accountModel/accountRequest.go
package accountModel

type CreateAccountRequest struct {
	Name        string `json:"name" validate:"required"`
	PhoneNumber string `json:"phone_number" validate:"required"`
	PIN         string `json:"pin" validate:"required"`
}